// src/core/types.ts
export type tooltipPlacementProp = 'top' | 'right' | 'bottom' | 'left' | 'auto';

export type tooltipTriggerProp = 'hover' | 'click';
// ...
export type colorProp =
  | 'primary'
  | 'secondary'
  | 'success'
  | 'info'
  | 'warning'
  | 'danger'
  | 'link'
  | 'black'
  | 'white'
  | 'light'
  | 'dark';